detangle scales
